requirements.txt
tqdm
numpy
numba==0.51.0
transforms3d
ruamel.yaml
torch
ignite
matplotlib
argparse
datetime
pioneer-common>=1.1
pioneer-das-api>=1.1